# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Senac Bike Store API - By Condé",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "INSIRA AQUI A URL do Host",
            "description": "Produção"
        },
        {
            "url": "INSIRA AQUI A URL do Host",
            "description": "Produção"
        }
    ],
    "paths": {
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Retorna a Lista de Categorias.",
                "description": "Retorna o JSON da Lista de Categorias.",
                "operationId": "getCategoriasList",
                "parameters": [
                    {
                        "name": "ordenacao",
                        "in": "query",
                        "description": "Ordernação dos dados",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriasResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriasResponseException500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Armazena os dados de uma nova Categoria.",
                "description": "Armazena e retorna o JSON de uma nova Categoria.",
                "operationId": "postCategoria",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoriaNewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException500"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Retorna os dados de uma Categoria.",
                "description": "Retorna o JSON de uma Categoria.",
                "operationId": "getCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Apaga os dados de uma categoria.",
                "description": "Apaga e retorna o JSON de confirmação.",
                "operationId": "deleteCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualiza os dados de uma categoria.",
                "description": "Atualiza e retorna o JSON de uma nova Categoria.",
                "operationId": "patchCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categoria"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoriaNewRequest": {
                "title": "CategoriaNewRequest",
                "description": "Requisição enviada via Body para uma inserção ou atualização de uma Categoria",
                "required": [
                    "nome_da_categoria"
                ],
                "properties": {
                    "nome_da_categoria": {
                        "title": "nome_da_categoria",
                        "description": "Nome da categoria.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CategoriaResponse": {
                "title": "CategoriaResponse",
                "description": "Resposta enviada via JSON para os dados de uma Categoria",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Categoria retornada."
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CategoriaResponse"
                }
            },
...
            "CategoriaResponseException406": {
                "title": "CategoriaResponseException406",
                "description": "Resposta enviada via JSON para os dados de uma Categoria",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 406
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "O campo Id deve ser numérico."
                    },
                    "categoria": {
                        "title": "categoria",
                        "description": "Dados de uma categoria.",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CategoriaResponseException406"
                }
            },
            "CategoriaResponseException500": {
                "title": "CategoriaResponseException500",
                "description": "Resposta enviada via JSON para os dados de uma Categoria",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 500
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Erro Interno."
                    },
                    "categoria": {
                        "title": "categoria",
                        "description": "Dados de uma categoria.",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CategoriaResponseException500"
                }
            },
            "CategoriasResponse": {
                "title": "CategoriasResponse",
                "description": "Resposta enviada via JSON para Lista de Categorias",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Lista de categorias retornada."
                    },
                    "categorias": {
                        "title": "categorias",
                        "description": "Lista de Categorias retornada.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Categoria"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CategoriasResponse"
                }
            },
            "CategoriasResponseException500": {
                "title": "CategoriasResponseException500",
                "description": "Resposta enviada via JSON para Lista de Categorias",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 500
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Erro Interno."
                    },
                    "categorias": {
                        "title": "categorias",
                        "description": "Lista de Categorias retornada.",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CategoriasResponseException500"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    }
}